backend/requirements.txt
fastapi==0.115.6
mysql_connector_repackaged==0.3.1
pdfplumber==0.11.5
pymysql==1.1.1
python-dotenv==1.0.1
selenium==4.28.0
SQLAlchemy==2.0.37
uvicorn==0.34.0
